Move test setup code into tools
